feat(vue-css-loader): change dist file name to be parsed by vue-loader
