+++ packages/hippy-vue-css-loader/package.json
@@ -2,7 +2,7 @@
   "name": "@hippy/vue-css-loader",
   "version": "unspecified",
   "description": "hippy-vue style loader module for webpack",
-  "main": "dist/index.js",
+  "main": "dist/css-loader.js",
   "homepage": "https://hippyjs.org",
   "repository": "https://github.com/Tencent/Hippy",
   "author": "OpenHippy Team",
